take mem_changer out from mem_writer
